refactor(work): extract runChatTurn from worker job
Move the agent execution loop out of apps/worker/src/jobs/work-run.ts
into a shared @neko/llm/work helper so the same loop can be invoked
from the Next.js process. The worker job becomes a thin pg-boss adapter:
it handles "stale job" detection (run row missing) and the per-job
seq counter, then delegates everything else.

Pure refactor — no behavior change. All emit calls, status messages,
backend.run options, DB writes, and error handling are preserved
verbatim. assistantText accumulation moves inside runChatTurn (where
the agent loop owns it) instead of leaking into the caller's emit
closure.

Test counts unchanged (54+29 passing); tsc clean on both worker and
web tsconfigs.

Sets up Phase 3 (in-process chat endpoint), which calls runChatTurn
with its own emit closure that writes to DB + pushes to SSE.
